hoàn thành bảng danh mục thức uống

+++ FinalProject/src/main/java/fp/coffeeshopmanagement/repository/DanhMucThucUongRepository.java
@@ -3,6 +3,6 @@ package fp.coffeeshopmanagement.repository;
 import org.springframework.data.repository.CrudRepository;
 import fp.coffeeshopmanagement.model.DanhMucThucUong;
 
-public interface DanhMucThucUongRepository extends CrudRepository<DanhMucThucUong, Long> {
+public interface DanhMucThucUongRepository extends CrudRepository<DanhMucThucUong, Integer> {
 
 }
