+++ FinalProject/src/main/java/fp/coffeeshopmanagement/service/DanhMucThucUongService.java
@@ -6,9 +6,9 @@ import fp.coffeeshopmanagement.model.*;
 public interface DanhMucThucUongService {
 	public List<DanhMucThucUong> getAllDanhMucThucUong();
 	 
-	 public DanhMucThucUong getDanhMucThucUongById(long id);
+	 public DanhMucThucUong getDanhMucThucUongById(int id);
 	 
 	 public void saveOrUpdate(DanhMucThucUong dmtu);
 	 
-	 public void deleteDanhMucThucUong(long id);
+	 public void deleteDanhMucThucUong(int id);
 }
